moved drawer to saparate component, App cleanup, allow App context
