+++ app/frontend/src/AppContext.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const AppContext = React.createContext();
+
+export default AppContext;
