+++ app/frontend/src/components/Menu/index.js
@@ -1,3 +1,3 @@
-export {default as MenuHeader} from './MenuHeader.jsx';
-export {default as MenuDrawer } from './MenuDrawer.jsx';
-export {default as Overlay} from './Overlay.jsx';
+export {default as Drawer } from './BaseDrawer.jsx';
+export {default as DrawerMenu } from './MenuDrawer.jsx';
+export {default as DrawerContent } from './ContentDrawer.jsx';
