+++ app/frontend/src/components/common.js
@@ -32,8 +32,8 @@ export const GlobalStyle = createGlobalStyle`
 `;
 
 export const Icon = styled.img`
-  width: 25px;
-  height: 25px;
+  width: 30px;
+  height: 30px;
 `;
 
 export const StyledIconButton = styled(IconButton)`
